Move classes into respective modules and fix tutorials
